(PC-24870)[API] refactor: move folder backoffice individual and collective booking in folder booking
