+++ api/src/pcapi/routes/backoffice/__init__.py
@@ -15,14 +15,14 @@ def install_routes(app: Flask) -> None:
     from .accounts import blueprint as accounts_blueprint
     from .admin import blueprint as admin_blueprint
     from .admin import bo_users_blueprint
-    from .collective_bookings import blueprint as collective_bookings_blueprint
+    from .bookings import collective_bookings_blueprint
+    from .bookings import individual_bookings_blueprint
     from .collective_offers import collective_offer_templates_blueprint
     from .collective_offers import collective_offers_blueprint
     from .criteria import blueprint as criteria_blueprint
     from .custom_reimbursement_rule import blueprint as custom_reimbursement_rule_blueprint
     from .finance import blueprint as finance_blueprint
     from .fraud import blueprint as fraud_blueprint
-    from .individual_bookings import blueprint as individual_bookings_blueprint
     from .multiple_offers import blueprint as multiple_offers_blueprint
     from .offer_validation_rules import blueprint as offer_validation_rule_blueprint
     from .offerers import offerer_blueprint
